Fixed bug in TreeItem component
Fixed a bug with `v-for` in the TreeItem component, where the key was
bound to the index of the task in the array, not the actual identity
of the task (the task ID). When deleting a task, this resulted in the
correct task being deleted on the server and in the client, but due to
Vue.js "in-place patch" strategy, the wrong child task was displayed,
because it took over another tasks index in the array, which was
wrongly used as the identity of a task.